добавил маппинг

+++ src/main/java/ru/skypro/homework/dto/ad/ExtendedAd.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class ExtendedAd {
-    private int pk;
+    private long pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
     private String email;
     private String image;
     private String phone;
-    private int price;
+    private long price;
     private String title;
 
 }
